refactored, trying to fix select_alt

+++ Source/usbd_audio.h
@@ -52,7 +52,7 @@ USBD_AUDIO_ControlTypeDef;
 
 typedef struct
 {
-  uint32_t                  tx_alt_setting, rx_alt_setting, midi_alt_setting; 
+  uint8_t                   tx_alt_setting, rx_alt_setting, midi_alt_setting;
 #ifdef USE_USBD_AUDIO_TX
   uint8_t                   audio_tx_buffer[AUDIO_TX_TOTAL_BUF_SIZE];
 #endif
